requirements.txt
numpy
scipy
sklearn
pyyaml
opencv-python
tensorflow
django